refactor - move setupBOM to same file
