+++ src/hello-world/public/locales/zh-CN/component.json
@@ -13,7 +13,7 @@
     "supplier-part-number": "Supplier Part Number",
     "order-qty": "Order Quantity",
     "rejection-rate": "Rejection Rate (%)",
-    "wastage-rate": "Wastage Rate (%)",
+    "scrap-rate": "Wastage Rate (%)",
     "revision": "Revision",
     "phase": "Phase",
     "category": "Category",
